API

Modules

Types and constants

Functions and macros

LegendTestData.legend_test_data_path

Documentation

LegendTestData.legend_test_data_path — Method

legend_test_data_path()::AbstractString

Get local path to the LEGEND test data set.

Will download the test data version matching the current version of LegendTestData if not already present. The local copy of the test data is managed via DataDeps.jl.

Set ENV["DATADEPS_ALWAYS_ACCEPT"] = "true" to avoid interactive prompt asking for download permission.